Create function_flagCount.R
When this data.frame is output to a csv, it is useful for the end-user to know how many non-missing values are there per column. This R function eliminates the additional tasks (filtering and counting in excel, etc.) the end-user will have to perform to find the number of non-missing values per column by displaying it at the top.